src/main/java/com/yu/sell/service/impl/CategroyServiceImpl.java
package com.yu.sell.service.impl;

import com.yu.sell.dataobject.ProductCategory;
import com.yu.sell.repository.ProductCategoryRepository;
import com.yu.sell.service.CategoryService;
import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * 类目
 * @author 75839
 */
@Service
@Slf4j
public class CategroyServiceImpl implements CategoryService {
    @Autowired
    private ProductCategoryRepository productCategoryRepository;

    @Override
    public ProductCategory findOne(Integer categoryId) {
        ProductCategory productCategory = (ProductCategory) productCategoryRepository.findById(categoryId).get();
        return productCategory;
    }

    @Override
    public List<ProductCategory> findAll() {
        return productCategoryRepository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return productCategoryRepository.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return productCategoryRepository.save(productCategory);
    }
}
